+++ core/econf/file/file.go
@@ -6,12 +6,14 @@ import (
 	"github.com/gotomicro/ego/core/econf/manager"
 	"io/ioutil"
 	"log"
+	"os"
 	"path/filepath"
+	"runtime"
+	"strings"
 
 	"github.com/BurntSushi/toml"
 	"github.com/fsnotify/fsnotify"
 	"github.com/gotomicro/ego/core/elog"
-	"github.com/gotomicro/ego/core/util/xfile"
 	"github.com/gotomicro/ego/core/util/xgo"
 )
 
@@ -51,7 +53,7 @@ func (fp *fileDataSource) Parse(path string, watch bool) {
 	if err != nil {
 		elog.Panic("new datasource", elog.Any("err", err))
 	}
-	dir := xfile.CheckAndGetParentDir(absolutePath)
+	dir := checkAndGetParentDir(absolutePath)
 	fp.path = absolutePath
 	fp.dir = dir
 	fp.enableWatch = watch
@@ -118,4 +120,45 @@ func (fp *fileDataSource) watch() {
 		log.Fatal(err)
 	}
 	<-done
+}
+
+// CheckAndGetParentDir ...
+func checkAndGetParentDir(path string) string {
+	// check path is the directory
+	isDir, err := isDirectory(path)
+	if err != nil || isDir {
+		return path
+	}
+	return getParentDirectory(path)
+}
+
+// IsDirectory ...
+func isDirectory(path string) (bool, error) {
+	f, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	switch mode := f.Mode(); {
+	case mode.IsDir():
+		return true, nil
+	case mode.IsRegular():
+		return false, nil
+	}
+	return false, nil
+}
+
+func getParentDirectory(dirctory string) string {
+	if runtime.GOOS == "windows" {
+		dirctory = strings.Replace(dirctory, "\\", "/", -1)
+	}
+	return substr(dirctory, 0, strings.LastIndex(dirctory, "/"))
+}
+
+func substr(s string, pos, length int) string {
+	runes := []rune(s)
+	l := pos + length
+	if l > len(runes) {
+		l = len(runes)
+	}
+	return string(runes[pos:l])
 }
