+++ core/util/xgo/goroutine.go
@@ -1,79 +1,20 @@
 package xgo
 
-import (
-	"sync"
-	"time"
-)
-
-// Serial 串行
-func Serial(fns ...func()) func() {
-	return func() {
-		for _, fn := range fns {
-			fn()
-		}
-	}
-}
-
-// Parallel 并发执行
-func Parallel(fns ...func()) func() {
-	var wg sync.WaitGroup
-	return func() {
-		wg.Add(len(fns))
-		for _, fn := range fns {
-			go fn()
-			wg.Done()
-			//go try2(fn, wg.Done)
-		}
-		wg.Wait()
-	}
-}
-
-// RestrictParallel 并发,最大并发量restrict
-func RestrictParallel(restrict int, fns ...func()) func() {
-	var channel = make(chan struct{}, restrict)
-	return func() {
-		var wg sync.WaitGroup
-		for _, fn := range fns {
-			wg.Add(1)
-			go func(fn func()) {
-				defer wg.Done()
-				channel <- struct{}{}
-				tryIgnoreFnReturn(fn, nil)
-				<-channel
-			}(fn)
-		}
-		wg.Wait()
-		close(channel)
-	}
-}
+import "golang.org/x/sync/errgroup"
 
 // Go goroutine
 func Go(fn func()) {
-	// 导致报错无法吐出
-	//go try2(fn, nil)
 	go fn()
 }
 
-// DelayGo goroutine
-func DelayGo(delay time.Duration, fn func()) {
-	//_, file, line, _ := runtime.Caller(1)
-	go func() {
-		//defer func() {
-		//	if err := recover(); err != nil {
-		//		_logger.err("recover", xlog.Any("err", err), xlog.String("line", fmt.Sprintf("%s:%d", file, line)))
-		//	}
-		//}()
-		time.Sleep(delay)
-		fn()
-	}()
-}
-
-// SafeGo safe go
-func SafeGo(fn func(), rec func(error)) {
-	go func() {
-		err := tryIgnoreFnReturn(fn, nil)
-		if err != nil {
-			rec(err)
+// ParallelWithError ...
+func ParallelWithError(fns ...func() error) func() error {
+	return func() error {
+		eg := errgroup.Group{}
+		for _, fn := range fns {
+			eg.Go(fn)
 		}
-	}()
+
+		return eg.Wait()
+	}
 }
