--- core/util/xstring/format.go
@@ -1,11 +1,0 @@
-package xstring
-
-import "fmt"
-
-// Formatter ...
-type Formatter string
-
-// Format ...
-func (fm Formatter) Format(args ...interface{}) string {
-	return fmt.Sprintf(string(fm), args...)
-}
